Remove `o.s.u.ui.SWTUtils`
This is a duplicate of `o.s.u.eclipse.SWTUtils`.

Fixes #1002288
